Feat/restructuring (#422)
* Did restructuring and made repo validation interceptor an optional bean instead as it makes it more clean

* Moved construction of FHIR servlet into a bean for better reuse of others that would like to depend directly on this library

* Disabled default validation enabled
