--- src/main/java/ca/uhn/fhir/jpa/starter/StarterJpaConfig.java
@@ -1,128 +1,0 @@
-package ca.uhn.fhir.jpa.starter;
-
-import ca.uhn.fhir.context.ConfigurationException;
-import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.jpa.api.IDaoRegistry;
-import ca.uhn.fhir.jpa.api.dao.IFhirSystemDao;
-import ca.uhn.fhir.jpa.batch.config.NonPersistedBatchConfigurer;
-import ca.uhn.fhir.jpa.config.util.HapiEntityManagerFactoryUtil;
-import ca.uhn.fhir.jpa.config.util.ResourceCountCacheUtil;
-import ca.uhn.fhir.jpa.config.util.ValidationSupportConfigUtil;
-import ca.uhn.fhir.jpa.dao.FulltextSearchSvcImpl;
-import ca.uhn.fhir.jpa.dao.IFulltextSearchSvc;
-import ca.uhn.fhir.jpa.dao.mdm.MdmLinkDaoJpaImpl;
-import ca.uhn.fhir.jpa.dao.search.HSearchSortHelperImpl;
-import ca.uhn.fhir.jpa.dao.search.IHSearchSortHelper;
-import ca.uhn.fhir.jpa.provider.DaoRegistryResourceSupportedSvc;
-import ca.uhn.fhir.jpa.search.DatabaseBackedPagingProvider;
-import ca.uhn.fhir.jpa.search.IStaleSearchDeletingSvc;
-import ca.uhn.fhir.jpa.search.StaleSearchDeletingSvcImpl;
-import ca.uhn.fhir.jpa.util.ResourceCountCache;
-import ca.uhn.fhir.jpa.validation.JpaValidationSupportChain;
-import ca.uhn.fhir.mdm.dao.IMdmLinkDao;
-import ca.uhn.fhir.rest.api.IResourceSupportedSvc;
-import ca.uhn.fhir.rest.server.util.ISearchParamRegistry;
-
-import org.hl7.fhir.common.hapi.validation.support.CachingValidationSupport;
-import org.springframework.batch.core.configuration.annotation.BatchConfigurer;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.core.env.ConfigurableEnvironment;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-
-import javax.persistence.EntityManagerFactory;
-import javax.sql.DataSource;
-
-@Configuration
-public class StarterJpaConfig {
-	@Bean
-	public IFulltextSearchSvc fullTextSearchSvc() {
-		return new FulltextSearchSvcImpl();
-	}
-
-	@Bean
-	public IStaleSearchDeletingSvc staleSearchDeletingSvc() {
-		return new StaleSearchDeletingSvcImpl();
-	}
-
-	@Primary
-	@Bean
-	public CachingValidationSupport validationSupportChain(JpaValidationSupportChain theJpaValidationSupportChain) {
-		return ValidationSupportConfigUtil.newCachingValidationSupport(theJpaValidationSupportChain);
-	}
-
-	@Bean
-	public BatchConfigurer batchConfigurer() {
-		return new NonPersistedBatchConfigurer();
-	}
-
-	@Autowired
-	AppProperties appProperties;
-	@Autowired
-	private DataSource myDataSource;
-	@Autowired
-	private ConfigurableEnvironment configurableEnvironment;
-
-	/**
-	 * Customize the default/max page sizes for search results. You can set these however
-	 * you want, although very large page sizes will require a lot of RAM.
-	 */
-	@Bean
-	public DatabaseBackedPagingProvider databaseBackedPagingProvider() {
-		DatabaseBackedPagingProvider pagingProvider = new DatabaseBackedPagingProvider();
-		pagingProvider.setDefaultPageSize(appProperties.getDefault_page_size());
-		pagingProvider.setMaximumPageSize(appProperties.getMax_page_size());
-		return pagingProvider;
-	}
-
-	@Bean
-	public IResourceSupportedSvc resourceSupportedSvc(IDaoRegistry theDaoRegistry) {
-		return new DaoRegistryResourceSupportedSvc(theDaoRegistry);
-	}
-
-	@Bean(name = "myResourceCountsCache")
-	public ResourceCountCache resourceCountsCache(IFhirSystemDao<?, ?> theSystemDao) {
-		return ResourceCountCacheUtil.newResourceCountCache(theSystemDao);
-	}
-
-	@Primary
-	@Bean
-	public LocalContainerEntityManagerFactoryBean entityManagerFactory(
-		ConfigurableListableBeanFactory myConfigurableListableBeanFactory, FhirContext theFhirContext) {
-		LocalContainerEntityManagerFactoryBean retVal = HapiEntityManagerFactoryUtil.newEntityManagerFactory(myConfigurableListableBeanFactory, theFhirContext);
-		retVal.setPersistenceUnitName("HAPI_PU");
-
-		try {
-			retVal.setDataSource(myDataSource);
-		} catch (Exception e) {
-			throw new ConfigurationException("Could not set the data source due to a configuration issue", e);
-		}
-		retVal.setJpaProperties(EnvironmentHelper.getHibernateProperties(configurableEnvironment, myConfigurableListableBeanFactory));
-		return retVal;
-	}
-
-	@Bean
-	@Primary
-	public JpaTransactionManager hapiTransactionManager(EntityManagerFactory entityManagerFactory) {
-		JpaTransactionManager retVal = new JpaTransactionManager();
-		retVal.setEntityManagerFactory(entityManagerFactory);
-		return retVal;
-	}
-
-	@Autowired
-	private ISearchParamRegistry mySearchParamRegistry;
-
-	@Bean
-	public IHSearchSortHelper hSearchSortHelper() {
-		return new HSearchSortHelperImpl(mySearchParamRegistry);
-	}
-
-	@Bean
-	public IMdmLinkDao<?> mdmLinkDao(){
-		return new MdmLinkDaoJpaImpl();
-	}
-}
