+++ src/main/java/ca/uhn/fhir/jpa/starter/common/FhirServerConfigR5.java
@@ -1,4 +1,4 @@
-package ca.uhn.fhir.jpa.starter;
+package ca.uhn.fhir.jpa.starter.common;
 
 import ca.uhn.fhir.jpa.config.r5.JpaR5Config;
 import ca.uhn.fhir.jpa.starter.annotations.OnR5Condition;
